Changed some functionnalyties in the map
--- The estimate is now properly filled when you type
the address.

--- Also, the infobox "Vous êtes ici" seems like it shouldn't be there
so I commented it out.

ALSO:: I think the estimate is not the good one so it should be checked.
How to do that estimate(I think) -->
Calculate the time for each taxi to the client / by the number of taxi.

Also, we might want to exclude taxi that are way to far from the client.
IE: don't include a taxi that is 800 km aways from the client..